requirements.txt
awscrt==0.23.0
beautifulsoup4==4.12.3
botocore==1.35.54
datasets==3.1.0
jmespath==1.0.1
matplotlib==3.9.2
networkx==3.4.2
numpy==2.1.3
PyYAML==6.0.2
Requests==2.32.3
s3transfer==0.10.3
setuptools==75.1.0
torch==2.5.1
transformers==4.46.2
